add health, armor and attack components, add logger

+++ Engine/TransformComponent.cpp
@@ -271,7 +271,7 @@ namespace XYZEngine
 		scale.x = std::sqrt(transform.GetMatrix()[0][0] * transform.GetMatrix()[0][0] + transform.GetMatrix()[1][0] * transform.GetMatrix()[1][0]);
 		scale.y = std::sqrt(transform.GetMatrix()[0][1] * transform.GetMatrix()[0][1] + transform.GetMatrix()[1][1] * transform.GetMatrix()[1][1]);
 
-		rotation = std::atan2(transform.GetMatrix()[0][1], transform.GetMatrix()[0][0]) * 180 / 3.14159265;
+		rotation = static_cast<float>(std::atan2(transform.GetMatrix()[0][1], transform.GetMatrix()[0][0])) * 180.f / 3.14159265f;
 	}
 	void TransformComponent::setLocalInfoFrom(const Matrix2D& transform) const
 	{
@@ -281,7 +281,7 @@ namespace XYZEngine
 		localScale.x = std::sqrt(transform.GetMatrix()[0][0] * transform.GetMatrix()[0][0] + transform.GetMatrix()[1][0] * transform.GetMatrix()[1][0]);
 		localScale.y = std::sqrt(transform.GetMatrix()[0][1] * transform.GetMatrix()[0][1] + transform.GetMatrix()[1][1] * transform.GetMatrix()[1][1]);
 
-		localRotation = std::atan2(transform.GetMatrix()[0][1], transform.GetMatrix()[0][0]) * 180 / 3.14159265;
+		localRotation = static_cast<float>(std::atan2(transform.GetMatrix()[0][1], transform.GetMatrix()[0][0])) * 180.f / 3.14159265f;
 	}
 	void TransformComponent::updateLocalTransform() const
 	{
